@@ -1,19 +1,24 @@
 import wget
 
-def wget_download(url,dir_file,desc=None):
+def wget_download(url, dir_file, desc=None):
     """
-    Download files by wget command
+    Downloads a file from a given URL using the wget library.
 
+    Usage:
+        >>> wget_out = wget_download(url, dir_file, desc)
     Inputs:
-        url -> [str] URL of the file to download
-        dir_file -> [str] Path of the file to store
-        desc -> [str,optional,default=None] Description of the downloading   
+        url -> [str] URL of the file to be downloaded.
+        dir_file -> [str] Path where the downloaded file will be stored.
+        desc -> [str, optional, default=None] Description of the download. If provided, it will be printed before the download starts.
     Outputs:
-        wget_out -> [str] Path of the file downloaded
-
+        wget_out -> [str] Path of the downloaded file.
     """
+
+    # Print the provided description
     if desc: print(desc)
-    wget_out = wget.download(url,dir_file)
-    print()
+
+    # Use wget to download the file and store it at the specified path
+    wget_out = wget.download(url, dir_file)
+    print()  # Print a newline for better output formatting
 
     return wget_out
